refactor(deck): correction des deck : 1 page par action + correction de la documentation des props pour les components
